+++ Source Files/Video.cpp
@@ -192,7 +192,7 @@ void Video::animateLeft(){
 
 			scaleFactor = scaleFactor - 0.01f;
 			videoFiles.at(firstVideoPosition).setThumbnailScale(scaleFactor, scaleFactor);
-			videoFiles.at(rightPosition).setThumbnailScale(1.5 - scaleFactor, 1.5 - scaleFactor);
+			videoFiles.at(rightPosition).setThumbnailScale((float)1.5 - scaleFactor, (float)1.5 - scaleFactor);
 
 			videoFiles.at(firstVideoPosition).moveThumbnail(0 - animationSpeed, 0);
 			videoFiles.at(rightPosition).moveThumbnail(0 - animationSpeed, 0);
@@ -215,7 +215,7 @@ void Video::animateRight() {
 
 			scaleFactor = scaleFactor - 0.01f;
 			videoFiles.at(firstVideoPosition).setThumbnailScale(scaleFactor, scaleFactor);
-			videoFiles.at(leftPosition).setThumbnailScale(1.5 - scaleFactor, 1.5 - scaleFactor);
+			videoFiles.at(leftPosition).setThumbnailScale((float)1.5 - scaleFactor, (float)1.5 - scaleFactor);
 
 			videoFiles.at(firstVideoPosition).moveThumbnail(animationSpeed, 0);
 			videoFiles.at(rightPosition).moveThumbnail(animationSpeed, 0);
